[bug] fix a bug in calculating last_access_dist, now both stack_dist and last_access_dist are same, the request at time t and t+1, the distance is 0 instead 1.

+++ libMimircache/cache/include/cacheUtils.h
@@ -21,7 +21,7 @@ typedef struct {
   guint64 n_exp_byte;
   guint64 n_total_obj;
   guint64 n_total_byte;
-  guint64 cur_time;
+  gint32 cur_time;
 } exp_info_t;
